Add Dependabot configuration
Dependabot is a service by GitHub that automatically scans the
dependencies of the project. If an outdated dependency is found,
Dependabot will create a PR for updating that dependency.

Add Dependabot configuration that will scan the GitHub workflows and Go
dependencies for possible updates.

Dependabot may have to be manually enabled on the Security tab of the
project.

More information: https://docs.github.com/en/code-security/dependabot