Add explicit flag to read from standard input
Fixes bug introduced in #240 where interactive use of `zk new` results
in the command sitting indefinitely for user to supply user input.

This change introduces an explicit flag to `new` to indicate standard
input should be read for the content of the note.  The flag is called
`-i` on `--input`.

The flag name is somewhat confusing as there is a `--no-input` flag at
the global level.  I propose that we change that flag name to
`--no-prompt` as it more accurately describes the option.  To maintain
backwards compatibility, we define `--no-input` as an alias.