feat(lsp): add 'zk.link' LSP command
Allows users to take advantage of the linking mechanism in their editing
tools outside of the 'zk.new' context. For example, I made a
"ZkInsertLink" command that inserts a link at the current cursor location
after searching for the desired note. This improves on the current
behavior which only autocompletes based on note name, which users may or
may not remember.