Fix mix-up in error message when aborting
If I abort editing a file, `zk` shows 'failed to launch editor'. This
message is not correct, since I explicitly aborted the operation after
the editor successfully launched.

When using Vim, this is triggered by using `:cq`, which exits non-zero.
With a GUI editor, this is usually achieved by pressing the Cancel/Abort
button.

Update the error handling to show the correct error: if the editor
executed and exited abnormally, indicate that the operation was aborted.
Otherwise, show the previous message.