fix: Ctrl+E does no longer create new notes in chooser
There was a --bind behavior change in fzf 0.53.0 which
could be causing the error.

execute action is for executing an external program in the alternate
screen without leaving fzf. In zk abort+execute was used
to close fzf and open a zk instance.

become action is similar to execute, but it replaces the
current fzf process with the specified command using execve(2) system
call.

This should have similar behavior to the previous abort+execute action.

https://github.com/junegunn/fzf/issues/3865
https://github.com/junegunn/fzf/blob/710ebdf9c14e0c88b0cfc843ce08edcdd4554571/man/man1/fzf.1#L1785-L1789
https://github.com/junegunn/fzf/commit/6ea38b44384e7a09a3863465dc3cc7b93cd7e781