Fix error parsing escaped backticks
I noticed bogus LSP diagnostics on a note indicating missing backlinks
for _some_ links, but those backlink were not absent. I moved the links
around the file in a binary-search style, and found that links after an
escaped backtick were being ignored.

My assumption was that parsing backticks was somehow broken. The tests
in this commit prove this hypothesis.

The code changes make these tests pass.